requirements.txt
langgraph==1.0.8
langchain==1.2.10
langchain-openai==1.1.9
langsmith==0.7.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
fastapi
uvicorn[standard]